+++ lib/api.ts
@@ -1,6 +1,12 @@
-const Request = require('./request');
-const { Region, RegionUS } = require('./regions');
-const { SendEmailRequest } = require('./api/requests');
+import Request, { BearerAuth, RequestData, RequestDefaults } from './request';
+import { Region, RegionUS } from './regions';
+import { SendEmailRequest } from './api/requests';
+
+type APIDefaults = RequestDefaults & { region: Region; url?: string };
+
+type Recipients = Record<string, unknown>;
+
+type BroadcastsAllowedRecipientFieldsKeys = keyof typeof BROADCASTS_ALLOWED_RECIPIENT_FIELDS;
 
 const BROADCASTS_ALLOWED_RECIPIENT_FIELDS = {
   ids: ['ids', 'id_ignore_missing'],
@@ -9,8 +15,8 @@ const BROADCASTS_ALLOWED_RECIPIENT_FIELDS = {
   data_file_url: ['data_file_url', 'id_ignore_missing', 'email_ignore_missing', 'email_add_duplicates'],
 };
 
-const filterRecipientsDataForField = (recipients, field) => {
-  return BROADCASTS_ALLOWED_RECIPIENT_FIELDS[field].reduce((obj, field) => {
+const filterRecipientsDataForField = (recipients: Recipients, field: BroadcastsAllowedRecipientFieldsKeys) => {
+  return BROADCASTS_ALLOWED_RECIPIENT_FIELDS[field].reduce<Record<string, unknown>>((obj, field) => {
     if (!!recipients[field]) {
       obj[field] = recipients[field];
     }
@@ -18,8 +24,13 @@ const filterRecipientsDataForField = (recipients, field) => {
   }, {});
 };
 
-class APIClient {
-  constructor(appKey, defaults = {}) {
+export class APIClient {
+  appKey: BearerAuth;
+  defaults: APIDefaults;
+  request: Request;
+  apiRoot: string;
+
+  constructor(appKey: BearerAuth, defaults: Partial<APIDefaults> = {}) {
     if (defaults.region && !(defaults.region instanceof Region)) {
       throw new Error('region must be one of Regions.US or Regions.EU');
     }
@@ -31,7 +42,7 @@ class APIClient {
     this.apiRoot = this.defaults.url ? this.defaults.url : this.defaults.region.apiUrl;
   }
 
-  sendEmail(req) {
+  sendEmail(req: SendEmailRequest) {
     if (!(req instanceof SendEmailRequest)) {
       throw new Error('"request" must be an instance of SendEmailRequest');
     }
@@ -39,9 +50,11 @@ class APIClient {
     return this.request.post(`${this.apiRoot}/send/email`, req.message);
   }
 
-  triggerBroadcast(id, data, recipients) {
+  triggerBroadcast(id: string | number, data: RequestData, recipients: Recipients) {
     let payload = {};
-    let customRecipientField = Object.keys(BROADCASTS_ALLOWED_RECIPIENT_FIELDS).find(field => recipients[field]);
+    let customRecipientField = (Object.keys(
+      BROADCASTS_ALLOWED_RECIPIENT_FIELDS,
+    ) as BroadcastsAllowedRecipientFieldsKeys[]).find((field) => recipients[field]);
 
     if (customRecipientField) {
       payload = Object.assign({ data }, filterRecipientsDataForField(recipients, customRecipientField));
@@ -56,7 +69,4 @@ class APIClient {
   }
 }
 
-module.exports = {
-  APIClient,
-  SendEmailRequest,
-};
+export { SendEmailRequest } from './api/requests';