+++ test/api.ts
@@ -1,7 +1,11 @@
-const test = require('ava');
-const sinon = require('sinon');
-const { APIClient, SendEmailRequest } = require('../api');
-const { RegionUS, RegionEU } = require('../lib/regions');
+import avaTest, { TestInterface } from 'ava';
+import sinon, { SinonStub } from 'sinon';
+import { APIClient, SendEmailRequest } from '../lib/api';
+import { RegionUS, RegionEU } from '../lib/regions';
+
+type TestContext = { client: APIClient };
+
+const test = avaTest as TestInterface<TestContext>;
 
 test.beforeEach((t) => {
   t.context.client = new APIClient('appKey');
@@ -23,10 +27,18 @@ test('constructor sets correct URL for different regions', (t) => {
   });
 });
 
+test('constructor sets correct URL for a custom URL', (t) => {
+  let client = new APIClient('appKey', { url: 'https://example.com' });
+
+  t.is(client.appKey, 'appKey');
+  t.truthy(client.request);
+  t.is(client.apiRoot, 'https://example.com');
+});
+
 test('passing in an invalid region throws an error', (t) => {
   t.throws(
     () => {
-      new APIClient('appKey', { region: 'au' });
+      new APIClient('appKey', { region: 'au' } as any);
     },
     {
       message: 'region must be one of Regions.US or Regions.EU',
@@ -36,23 +48,38 @@ test('passing in an invalid region throws an error', (t) => {
 
 test('sendEmail: passing in a plain object throws an error', (t) => {
   sinon.stub(t.context.client.request, 'post');
+
   let req = { identifiers: { id: '2' }, transactional_message_id: 1 };
-  t.throws(() => t.context.client.sendEmail(req), {
+
+  t.throws(() => t.context.client.sendEmail(req as any), {
     message: /"request" must be an instance of SendEmailRequest/,
   });
-  t.falsy(t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/send/email`));
+  t.falsy((t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/send/email`));
 });
 
-test('#sendEmail: success', (t) => {
+test('#sendEmail: with template: success', (t) => {
   sinon.stub(t.context.client.request, 'post');
-  let req = new SendEmailRequest({ identifiers: { id: '2' }, transactional_message_id: 1 });
+  let req = new SendEmailRequest({ to: 'test@example.com', identifiers: { id: '2' }, transactional_message_id: 1 });
   t.context.client.sendEmail(req);
-  t.truthy(t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/send/email`, req.message));
+  t.truthy((t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/send/email`, req.message));
+});
+
+test('#sendEmail: without template: success', (t) => {
+  sinon.stub(t.context.client.request, 'post');
+  let req = new SendEmailRequest({
+    to: 'test@example.com',
+    identifiers: { id: '2' },
+    from: 'admin@example.com',
+    subject: 'This is a test',
+    body: 'Hi there!',
+  });
+  t.context.client.sendEmail(req);
+  t.truthy((t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/send/email`, req.message));
 });
 
 test('#sendEmail: adding attachments with encoding (default)', (t) => {
   sinon.stub(t.context.client.request, 'post');
-  let req = new SendEmailRequest({ identifiers: { id: '2' }, transactional_message_id: 1 });
+  let req = new SendEmailRequest({ to: 'test@example.com', identifiers: { id: '2' }, transactional_message_id: 1 });
 
   req.attach('test', 'hello world');
   t.is(req.message.attachments.test, Buffer.from('hello world').toString('base64'));
@@ -60,7 +87,7 @@ test('#sendEmail: adding attachments with encoding (default)', (t) => {
 
 test('#sendEmail: adding attachments without encoding', (t) => {
   sinon.stub(t.context.client.request, 'post');
-  let req = new SendEmailRequest({ identifiers: { id: '2' }, transactional_message_id: 1 });
+  let req = new SendEmailRequest({ to: 'test@example.com', identifiers: { id: '2' }, transactional_message_id: 1 });
 
   req.attach('file', 'test content', { encode: false });
   t.truthy(req.message.attachments.file, 'test content');
@@ -68,7 +95,7 @@ test('#sendEmail: adding attachments without encoding', (t) => {
 
 test('#sendEmail: adding attachments twice throws an error', (t) => {
   sinon.stub(t.context.client.request, 'post');
-  let req = new SendEmailRequest({ identifiers: { id: '2' }, transactional_message_id: 1 });
+  let req = new SendEmailRequest({ to: 'test@example.com', identifiers: { id: '2' }, transactional_message_id: 1 });
 
   req.attach('test', 'test content');
   t.throws(() => req.attach('test', 'test content 2'), { message: /attachment test already exists/ });
@@ -78,20 +105,20 @@ test('#sendEmail: adding attachments twice throws an error', (t) => {
 test('#sendEmail: error', async (t) => {
   sinon.stub(t.context.client.request, 'post').rejects({ message: 'sample error', statusCode: 400 });
 
-  let req = new SendEmailRequest({ identifiers: { id: '2' }, transactional_message_id: 1 });
+  let req = new SendEmailRequest({ to: 'test@example.com', identifiers: { id: '2' }, transactional_message_id: 1 });
   t.context.client.sendEmail(req).catch((err) => {
     t.is(err.message, 'sample error');
     t.is(err.statusCode, 400);
   });
 
-  t.truthy(t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/send/email`, req.message));
+  t.truthy((t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/send/email`, req.message));
 });
 
 test('#triggerBroadcast works', (t) => {
   sinon.stub(t.context.client.request, 'post');
   t.context.client.triggerBroadcast(1, { type: 'data' }, { type: 'recipients' });
   t.truthy(
-    t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
+    (t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
       data: { type: 'data' },
       recipients: { type: 'recipients' },
     }),
@@ -110,7 +137,7 @@ test('#triggerBroadcast works with emails', (t) => {
     },
   );
   t.truthy(
-    t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
+    (t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
       data: { type: 'data' },
       emails: ['test@email.com'],
       email_ignore_missing: true,
@@ -123,7 +150,7 @@ test('#triggerBroadcast works with ids', (t) => {
   sinon.stub(t.context.client.request, 'post');
   t.context.client.triggerBroadcast(1, { type: 'data' }, { ids: [1], id_ignore_missing: true });
   t.truthy(
-    t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
+    (t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
       data: { type: 'data' },
       ids: [1],
       id_ignore_missing: true,
@@ -136,7 +163,7 @@ test('#triggerBroadcast works with per_user_data', (t) => {
   const per_user_data = [{ id: 1, data: { very: 'important' } }];
   t.context.client.triggerBroadcast(1, { type: 'data' }, { per_user_data, id_ignore_missing: true });
   t.truthy(
-    t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
+    (t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
       data: { type: 'data' },
       per_user_data,
       id_ignore_missing: true,
@@ -149,7 +176,7 @@ test('#triggerBroadcast works with data_file_url', (t) => {
   const data_file_url = 'https://my.s3.bucket.com';
   t.context.client.triggerBroadcast(1, { type: 'data' }, { data_file_url, id_ignore_missing: true });
   t.truthy(
-    t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
+    (t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
       data: { type: 'data' },
       data_file_url,
       id_ignore_missing: true,
@@ -170,7 +197,7 @@ test('#triggerBroadcast discards extraneous fields', (t) => {
     },
   );
   t.truthy(
-    t.context.client.request.post.calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
+    (t.context.client.request.post as SinonStub).calledWith(`${RegionUS.apiUrl}/api/campaigns/1/triggers`, {
       data: { type: 'data' },
       ids: [1],
       id_ignore_missing: true,